Update according to review comments
