+++ test_scripts/Security/GetSystemTime/001_GetSystemTime_is_sent.lua
@@ -6,16 +6,16 @@
 --
 -- Description:
 -- In case:
--- 1) Mobile app starts secure RPC service
--- 2) Mobile certificate is up to date and sdl certificates is not up to date
--- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile certificate is valid and sdl certificate becomes valid
+-- 1) HMI sends BC.OnSystemTimeReady notification to SDL during initialization
+-- 2) Mobile and sdl certificates are valid
+-- 3) Mobile app starts secure service
 -- SDL must:
--- 1) not trigger PTU
--- 2) start secure service, Handshake is finished with frameInfo = START_SERVICE_ACK, encryption = true
+-- 1) Request GetSystemTime from HMI
+-- 2) Mobile and SDL certificates are valid according to date/time from GetSystemTime response
+-- 3) Start secure service: Handshake is finished with frameInfo = START_SERVICE_ACK, encryption = true
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
@@ -28,24 +28,11 @@ local pData = {
   encryption = true
 }
 
-local systemTime = {
-  millisecond = 100,
-  second = 30,
-  minute = 29,
-  hour = 15,
-  day = 20,
-  month = 1,
-  year = 2023,
-  tz_hour = -3,
-  tz_minute = 10
-}
-
 --[[ Local Functions ]]
-local function ptUpdateWithNotActualCer(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
+local function ptUpdate(pTbl)
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
 end
 
 --[[ Scenario ]]
@@ -57,9 +44,8 @@ runner.Title("Test")
 
 runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
-runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdateWithNotActualCer })
-runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecured,
-	{ pData, serviceId, 1, systemTime })
+runner.Step("PolicyTableUpdate with certificate", common.policyTableUpdate, { ptUpdate })
+runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecured, { pData, serviceId, 1 })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
