+++ test_scripts/Security/GetSystemTime/003_GetSystemTime_mobile_sdl_cer_are_not_valid_yet.lua
@@ -9,13 +9,13 @@
 -- 1) Mobile app starts secure RPC service
 -- 2) Mobile and sdl certificates are not up to date yet
 -- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile and sdl certificates are not valid
+-- 4) Mobile and sdl certificates are not valid according to date/time from GetSystemTime response
 -- SDL must:
 -- 1) trigger PTU and receives one more time not up to date sdl certificate during update
--- 2) not start secure service, Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
+-- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ General configuration parameters ]]
@@ -50,7 +50,6 @@ local function ptUpdate(pTbl)
   local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
 end
 
 --[[ Scenario ]]
