+++ test_scripts/Security/GetSystemTime/005_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_is_not_valid.lua
@@ -9,13 +9,13 @@
 -- 1) Mobile app starts secure RPC service
 -- 2) Mobile and sdl certificates are expired
 -- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile is still not valid and sdl becomes valid
+-- 4) Mobile certificate becomes valid and sdl certificate is not valid according to date/time from GetSystemTime response
 -- SDL must:
--- 1) not trigger PTU
--- 2) not start secure service, Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
+-- 1) trigger PTU and receives one more time expired certificate during update
+-- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ General configuration parameters ]]
@@ -40,17 +40,16 @@ local systemTime = {
   hour = 15,
   day = 20,
   month = 1,
-  year = 2022,
+  year = 2025,
   tz_hour = -3,
   tz_minute = 10
 }
 
 --[[ Local Functions ]]
 local function ptUpdate(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0312_17.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
 end
 
 --[[ Scenario ]]
@@ -63,8 +62,8 @@ runner.Title("Test")
 runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
 runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdate })
-runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecured,
-	{ pData, serviceId, 1, systemTime })
+runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecuredwithPTU,
+	{ pData, serviceId, 1, systemTime, ptUpdate })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
