+++ test_scripts/Security/GetSystemTime/006_GetSystemTime_mobile_is_not_valid_sdl_cer_becomes_valid.lua
@@ -7,15 +7,15 @@
 -- Description:
 -- In case:
 -- 1) Mobile app starts secure RPC service
--- 2) Mobile certificate is not up to date and sdl certificates is up to date
+-- 2) Mobile and sdl certificates are expired
 -- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile certificate becomes valid and sdl certificate is still valid
+-- 4) Mobile is still not valid and sdl becomes valid according to date/time from GetSystemTime response
 -- SDL must:
 -- 1) not trigger PTU
--- 2) start secure service, Handshake is finished with frameInfo = START_SERVICE_ACK, encryption = true
+-- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ General configuration parameters ]]
@@ -29,8 +29,8 @@ runner.testSettings.isSelfIncluded = false
 --[[ Local Variables ]]
 local serviceId = 7
 local pData = {
-  frameInfo = common.frameInfo.START_SERVICE_ACK,
-  encryption = true
+  frameInfo = common.frameInfo.START_SERVICE_NACK,
+  encryption = false
 }
 
 local systemTime = {
@@ -40,17 +40,16 @@ local systemTime = {
   hour = 15,
   day = 20,
   month = 1,
-  year = 2025,
+  year = 2022,
   tz_hour = -3,
   tz_minute = 10
 }
 
 --[[ Local Functions ]]
 local function ptUpdate(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential.pem"
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
 end
 
 --[[ Scenario ]]
