+++ test_scripts/Security/GetSystemTime/009_GetSystemTime_mobile_sdl_cer_become_valid_after_PTU.lua
@@ -9,13 +9,13 @@
 -- 1) Mobile app starts secure RPC service
 -- 2) Mobile and sdl certificates are expired
 -- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile certificate becomes valid and sdl certificate is not valid
+-- 4) Mobile certificate becomes valid and sdl are still not valid according to date/time from GetSystemTime response
 -- SDL must:
--- 1) trigger PTU and receives one more time expired certificate during update
--- 2) not start secure service, Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
+-- 1) SDL triggers PTU and sdl certificate becomes valid
+-- 2) Start secure service: Handshake is finished with frameInfo = START_SERVICE_ACK, encryption = true
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ General configuration parameters ]]
@@ -29,8 +29,8 @@ runner.testSettings.isSelfIncluded = false
 --[[ Local Variables ]]
 local serviceId = 7
 local pData = {
-  frameInfo = common.frameInfo.START_SERVICE_NACK,
-  encryption = false
+  frameInfo = common.frameInfo.START_SERVICE_ACK,
+  encryption = true
 }
 
 local systemTime = {
@@ -40,17 +40,22 @@ local systemTime = {
   hour = 15,
   day = 20,
   month = 1,
-  year = 2025,
+  year = 2024,
   tz_hour = -3,
   tz_minute = 10
 }
 
 --[[ Local Functions ]]
-local function ptUpdate(pTbl)
+local function ptUpdateWithNotValidCer(pTbl)
   local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0312_17.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
+end
+
+local function ptUpdateWithValidCer(pTbl)
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
+  local crt = common.readFile(filePath)
+  pTbl.policy_table.module_config.certificate = crt
 end
 
 --[[ Scenario ]]
@@ -62,9 +67,9 @@ runner.Title("Test")
 
 runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
-runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdate })
+runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdateWithNotValidCer })
 runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecuredwithPTU,
-	{ pData, serviceId, 1, systemTime, ptUpdate })
+	{ pData, serviceId, 1, systemTime, ptUpdateWithValidCer })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
