+++ test_scripts/Security/GetSystemTime/010_GetSystemTime_mobile_cer_expired_sdl_cer_valid.lua
@@ -9,13 +9,13 @@
 -- 1) Mobile app starts secure RPC service
 -- 2) Mobile certificate is expired and sdl certificates is up to date
 -- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile certificate is still not valid and sdl certificate is valid
+-- 4) Mobile certificate is still not valid and sdl certificate is valid according to date/time from GetSystemTime response
 -- SDL must:
 -- 1) not trigger PTU
--- 2) not start secure service, Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
+-- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ General configuration parameters ]]
@@ -50,7 +50,6 @@ local function ptUpdate(pTbl)
   local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0316_21.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
 end
 
 --[[ Scenario ]]
