+++ test_scripts/Security/GetSystemTime/011_GetSystemTime_mobile_cer_not_valid_yet_sdl_cer_valid.lua
@@ -7,15 +7,15 @@
 -- Description:
 -- In case:
 -- 1) Mobile app starts secure RPC service
--- 2) Mobile and sdl certificates are expired
+-- 2) Mobile certificate is not up to date yet and sdl certificates is up to date
 -- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile and sdl are still not valid
+-- 4) Mobile certificate is still not valid and sdl certificates is valid according to date/time from GetSystemTime response
 -- SDL must:
--- 1) SDL triggers PTU and sdl certificate becomes valid
--- 2) not start secure service, Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
+-- 1) not trigger PTU
+-- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ General configuration parameters ]]
@@ -40,24 +40,16 @@ local systemTime = {
   hour = 15,
   day = 20,
   month = 1,
-  year = 2018,
+  year = 2020,
   tz_hour = -3,
   tz_minute = 10
 }
 
 --[[ Local Functions ]]
-local function ptUpdateWithNotValidCer(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
-  local crt = common.readFile(filePath)
-  pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
-end
-
-local function ptUpdateWithValidCer(pTbl)
+local function ptUpdate(pTbl)
   local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0316_21.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
 end
 
 --[[ Scenario ]]
@@ -69,9 +61,9 @@ runner.Title("Test")
 
 runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
-runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdateWithNotValidCer })
-runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecuredwithPTU,
-	{ pData, serviceId, 1, systemTime, ptUpdateWithValidCer })
+runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdate })
+runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecured,
+	{ pData, serviceId, 1, systemTime })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
