+++ test_scripts/Security/GetSystemTime/014_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_becomes_not_valid.lua
@@ -7,15 +7,15 @@
 -- Description:
 -- In case:
 -- 1) Mobile app starts secure RPC service
--- 2) Mobile and sdl certificates are expired
+-- 2) Mobile certificate is not up to date and sdl certificates is up to date
 -- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile certificate becomes valid and sdl are still not valid
+-- 4) Mobile certificate becomes valid and sdl certificate becomes not valid according to date/time from GetSystemTime response
 -- SDL must:
--- 1) SDL triggers PTU and sdl certificate becomes valid
--- 2) start secure service, Handshake is finished with frameInfo = START_SERVICE_ACK, encryption = true
+-- 1) trigger PTU
+-- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ General configuration parameters ]]
@@ -29,8 +29,8 @@ runner.testSettings.isSelfIncluded = false
 --[[ Local Variables ]]
 local serviceId = 7
 local pData = {
-  frameInfo = common.frameInfo.START_SERVICE_ACK,
-  encryption = true
+  frameInfo = common.frameInfo.START_SERVICE_NACK,
+  encryption = false
 }
 
 local systemTime = {
@@ -40,24 +40,16 @@ local systemTime = {
   hour = 15,
   day = 20,
   month = 1,
-  year = 2024,
+  year = 2027,
   tz_hour = -3,
   tz_minute = 10
 }
 
 --[[ Local Functions ]]
-local function ptUpdateWithNotValidCer(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0312_17.pem"
+local function ptUpdate(pTbl)
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
-end
-
-local function ptUpdateWithValidCer(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
-  local crt = common.readFile(filePath)
-  pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
 end
 
 --[[ Scenario ]]
@@ -69,9 +61,9 @@ runner.Title("Test")
 
 runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
-runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdateWithNotValidCer })
+runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdate })
 runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecuredwithPTU,
-	{ pData, serviceId, 1, systemTime, ptUpdateWithValidCer })
+	{ pData, serviceId, 1, systemTime, ptUpdate })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
