+++ test_scripts/Security/GetSystemTime/015_GetSystemTime_mobile_cer_valid_sdl_cer_not_valid.lua
@@ -7,24 +7,19 @@
 -- Description:
 -- In case:
 -- 1) Mobile app starts secure RPC service
--- 2) Mobile and sdl certificates are up to date
+-- 2) Mobile certificate is up to date and sdl certificates is not up to date
 -- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile certificate becomes not valid yet and sdl certificate is still valid
+-- 4) Mobile certificate is valid and sdl certificate is not valid yet according to date/time from GetSystemTime response
 -- SDL must:
--- 1) not trigger PTU
--- 2) not start secure service, Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
+-- 1) trigger PTU
+-- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
-
---[[ General configuration parameters ]]
-config.serverCertificatePath = "./files/Security/GetSystemTime_certificates/spt_credential_0317_22.pem"
-config.serverPrivateKeyPath = "./files/Security/GetSystemTime_certificates/spt_credential_0317_22.pem"
-config.serverCAChainCertPath = "./files/Security/GetSystemTime_certificates/spt_credential_0317_22.pem"
 
 --[[ Local Variables ]]
 local serviceId = 7
@@ -40,17 +35,22 @@ local systemTime = {
   hour = 15,
   day = 20,
   month = 1,
-  year = 2015,
+  year = 2020,
   tz_hour = -3,
   tz_minute = 10
 }
 
 --[[ Local Functions ]]
 local function ptUpdateWithNotActualCer(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential.pem"
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
+end
+
+local function ptUpdateWithExpiredCer(pTbl)
+  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0312_17.pem"
+  local crt = common.readFile(filePath)
+  pTbl.policy_table.module_config.certificate = crt
 end
 
 --[[ Scenario ]]
@@ -63,8 +63,8 @@ runner.Title("Test")
 runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
 runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdateWithNotActualCer })
-runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecured,
-	{ pData, serviceId, 1, systemTime })
+runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecuredwithPTU,
+	{ pData, serviceId, 1, systemTime, ptUpdateWithExpiredCer })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
