+++ test_scripts/Security/GetSystemTime/025_GetSystemTime_mobile_cer_and_sdl_cer_become_expired.lua
@@ -6,16 +6,16 @@
 --
 -- Description:
 -- In case:
--- 1) HMI sends BC.OnSystemTimeReady notification to SDL during initialization
--- 2) Mobile and sdl certificates are valid
--- 3) Mobile app starts secure service
+-- 1) Mobile app starts secure RPC service
+-- 2) Mobile and sdl certificates are up to date
+-- 3) SDL requests GetSystemTime
+-- 4) Mobile certificate and sdl certificate become expired according to date/time from GetSystemTime response
 -- SDL must:
--- 1) SDL requests GetSystemTime to HMI
--- 2) According to time from GetSystemTime response mobile and SDL certificates are valid
--- 4) start secure service, Handshake is finished with frameInfo = START_SERVICE_ACK, encryption = true
+-- 1) trigger PTU
+-- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
@@ -24,16 +24,27 @@ runner.testSettings.isSelfIncluded = false
 --[[ Local Variables ]]
 local serviceId = 7
 local pData = {
-  frameInfo = common.frameInfo.START_SERVICE_ACK,
-  encryption = true
+  frameInfo = common.frameInfo.START_SERVICE_NACK,
+  encryption = false
+}
+
+local systemTime = {
+  millisecond = 100,
+  second = 30,
+  minute = 29,
+  hour = 15,
+  day = 20,
+  month = 1,
+  year = 2028,
+  tz_hour = -3,
+  tz_minute = 10
 }
 
 --[[ Local Functions ]]
-local function ptUpdate(pTbl)
+local function ptUpdateWithNotActualCer(pTbl)
   local filePath = "./files/Security/GetSystemTime_certificates/client_credential.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
 end
 
 --[[ Scenario ]]
@@ -45,8 +56,9 @@ runner.Title("Test")
 
 runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
-runner.Step("PolicyTableUpdate with certificate", common.policyTableUpdate, { ptUpdate })
-runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecured, { pData, serviceId, 1 })
+runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdateWithNotActualCer })
+runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecuredwithPTU,
+	{ pData, serviceId, 1, systemTime, ptUpdateWithNotActualCer })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
