+++ test_scripts/Security/GetSystemTime/026_GetSystemTime_mobile_cer_and_sdl_cer_become_not_valid_yet.lua
@@ -7,15 +7,15 @@
 -- Description:
 -- In case:
 -- 1) Mobile app starts secure RPC service
--- 2) Mobile certificate is up to date and sdl certificates is not up to date
+-- 2) Mobile and sdl certificates are up to date
 -- 3) SDL requests GetSystemTime
--- 4) According to time from GetSystemTime response mobile certificate is valid and sdl certificate is not valid
+-- 4) Mobile certificate and sdl certificate become not valid yet according to date/time from GetSystemTime response
 -- SDL must:
--- 1) trigger PTU and sdl certificate becomes valid
--- 2) start secure service, Handshake is finished with frameInfo = START_SERVICE_ACK, encryption = true
+-- 1) trigger PTU
+-- 2) Not start secure service: Handshake is finished with frameInfo = START_SERVICE_NACK, encryption = false
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/GetSystemTime/common')
+local common = require('test_scripts/Security/GetSystemTime/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
@@ -24,8 +24,8 @@ runner.testSettings.isSelfIncluded = false
 --[[ Local Variables ]]
 local serviceId = 7
 local pData = {
-  frameInfo = common.frameInfo.START_SERVICE_ACK,
-  encryption = true
+  frameInfo = common.frameInfo.START_SERVICE_NACK,
+  encryption = false
 }
 
 local systemTime = {
@@ -35,24 +35,16 @@ local systemTime = {
   hour = 15,
   day = 20,
   month = 1,
-  year = 2020,
+  year = 2011,
   tz_hour = -3,
   tz_minute = 10
 }
 
 --[[ Local Functions ]]
 local function ptUpdateWithNotActualCer(pTbl)
-  local filePath = "./files/Security/GetSystemTime_certificates/client_credential_0321_26.pem"
-  local crt = common.readFile(filePath)
-  pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
-end
-
-local function ptUpdateWithActualCer(pTbl)
   local filePath = "./files/Security/GetSystemTime_certificates/client_credential.pem"
   local crt = common.readFile(filePath)
   pTbl.policy_table.module_config.certificate = crt
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { common.appHMIType }
 end
 
 --[[ Scenario ]]
@@ -66,7 +58,7 @@ runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
 runner.Step("PolicyTableUpdate with not valid certificate", common.policyTableUpdate, { ptUpdateWithNotActualCer })
 runner.Step("Handshake with BC.GetSystemTime request from SDL", common.startServiceSecuredwithPTU,
-	{ pData, serviceId, 1, systemTime, ptUpdateWithActualCer })
+	{ pData, serviceId, 1, systemTime, ptUpdateWithNotActualCer })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
