+++ test_sets/get_system_time.txt
@@ -1,26 +1,27 @@
-./test_scripts/Policies/GetSystemTime/001_GetSystemTime_is_sent.lua
-./test_scripts/Policies/GetSystemTime/002_GetSystemTime_is_not_sent.lua
-./test_scripts/Policies/GetSystemTime/003_GetSystemTime_mobile_sdl_cer_are_not_valid_yet.lua
-./test_scripts/Policies/GetSystemTime/004_GetSystemTime_mobile_sdl_cer_are_not_valid_already.lua
-./test_scripts/Policies/GetSystemTime/005_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_is_not_valid.lua
-./test_scripts/Policies/GetSystemTime/006_GetSystemTime_mobile_is_not_valid_sdl_cer_becomes_valid.lua
-./test_scripts/Policies/GetSystemTime/007_GetSystemTime_mobile_is_not_valid_sdl_cer_becomes_valid_after_PTU.lua
-./test_scripts/Policies/GetSystemTime/008_GetSystemTime_mobile_sdl_cer_become_valid.lua
-./test_scripts/Policies/GetSystemTime/009_GetSystemTime_mobile_sdl_cer_become_valid_after_PTU.lua
-./test_scripts/Policies/GetSystemTime/010_GetSystemTime_mobile_cer_expired_sdl_cer_valid.lua
-./test_scripts/Policies/GetSystemTime/011_GetSystemTime_mobile_cer_not_valid_yet_sdl_cer_valid.lua
-./test_scripts/Policies/GetSystemTime/012_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_valid.lua
-./test_scripts/Policies/GetSystemTime/013_GetSystemTime_mobile_cer_not_valid_sdl_becomes_invalid.lua
-./test_scripts/Policies/GetSystemTime/014_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_becomes_not_valid.lua
-./test_scripts/Policies/GetSystemTime/015_GetSystemTime_mobile_cer_valid_sdl_cer_not_valid.lua
-./test_scripts/Policies/GetSystemTime/016_GetSystemTime_mobile_cer_valid_sdl_cer_becomes_valid.lua
-./test_scripts/Policies/GetSystemTime/017_GetSystemTime_mobile_cer_valid_sdl_cer_becomes_valid_after_PTU.lua
-./test_scripts/Policies/GetSystemTime/018_GetSystemTime_mobile_cer_becomes_not_valid_sdl_cer_not_valid.lua
-./test_scripts/Policies/GetSystemTime/019_GetSystemTime_mobile_cer_becomes_not_valid_sdl_cer_becomes_valid.lua
-./test_scripts/Policies/GetSystemTime/020_GetSystemTime_mobile_cer_becomes_not_valid_sdl_cer_becomes_valid_with_PTU.lua
-./test_scripts/Policies/GetSystemTime/021_GetSystemTime_mobile_cer_and_sdl_cer_valid.lua
-./test_scripts/Policies/GetSystemTime/022_GetSystemTime_mobile_cer_becomes_not_valid_and_sdl_cer_valid.lua
-./test_scripts/Policies/GetSystemTime/023_GetSystemTime_mobile_cer_becomes_yet_not_valid_and_sdl_cer_valid.lua
-./test_scripts/Policies/GetSystemTime/024_GetSystemTime_mobile_cer_valid_and_sdl_cer_becomes_not_valid.lua
-./test_scripts/Policies/GetSystemTime/025_GetSystemTime_mobile_cer_and_sdl_cer_become_expired.lua
-./test_scripts/Policies/GetSystemTime/026_GetSystemTime_mobile_cer_and_sdl_cer_become_not_valid_yet.lua
+./test_scripts/Security/GetSystemTime/001_GetSystemTime_is_sent.lua
+./test_scripts/Security/GetSystemTime/002_GetSystemTime_is_not_sent.lua
+./test_scripts/Security/GetSystemTime/003_GetSystemTime_mobile_sdl_cer_are_not_valid_yet.lua
+./test_scripts/Security/GetSystemTime/004_GetSystemTime_mobile_sdl_cer_are_not_valid_already.lua
+./test_scripts/Security/GetSystemTime/005_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_is_not_valid.lua
+./test_scripts/Security/GetSystemTime/006_GetSystemTime_mobile_is_not_valid_sdl_cer_becomes_valid.lua
+./test_scripts/Security/GetSystemTime/007_GetSystemTime_mobile_is_not_valid_sdl_cer_becomes_valid_after_PTU.lua
+./test_scripts/Security/GetSystemTime/008_GetSystemTime_mobile_sdl_cer_become_valid.lua
+./test_scripts/Security/GetSystemTime/009_GetSystemTime_mobile_sdl_cer_become_valid_after_PTU.lua
+./test_scripts/Security/GetSystemTime/010_GetSystemTime_mobile_cer_expired_sdl_cer_valid.lua
+./test_scripts/Security/GetSystemTime/011_GetSystemTime_mobile_cer_not_valid_yet_sdl_cer_valid.lua
+./test_scripts/Security/GetSystemTime/012_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_valid.lua
+./test_scripts/Security/GetSystemTime/013_GetSystemTime_mobile_cer_not_valid_sdl_becomes_invalid.lua
+./test_scripts/Security/GetSystemTime/014_GetSystemTime_mobile_cer_becomes_valid_sdl_cer_becomes_not_valid.lua
+./test_scripts/Security/GetSystemTime/015_GetSystemTime_mobile_cer_valid_sdl_cer_not_valid.lua
+./test_scripts/Security/GetSystemTime/016_GetSystemTime_mobile_cer_valid_sdl_cer_becomes_valid.lua
+./test_scripts/Security/GetSystemTime/017_GetSystemTime_mobile_cer_valid_sdl_cer_becomes_valid_after_PTU.lua
+./test_scripts/Security/GetSystemTime/018_GetSystemTime_mobile_cer_becomes_not_valid_sdl_cer_not_valid.lua
+./test_scripts/Security/GetSystemTime/019_GetSystemTime_mobile_cer_becomes_not_valid_sdl_cer_becomes_valid.lua
+./test_scripts/Security/GetSystemTime/020_GetSystemTime_mobile_cer_becomes_not_valid_sdl_cer_becomes_valid_with_PTU.lua
+./test_scripts/Security/GetSystemTime/021_GetSystemTime_mobile_cer_and_sdl_cer_valid.lua
+./test_scripts/Security/GetSystemTime/022_GetSystemTime_mobile_cer_becomes_not_valid_and_sdl_cer_valid.lua
+./test_scripts/Security/GetSystemTime/023_GetSystemTime_mobile_cer_becomes_yet_not_valid_and_sdl_cer_valid.lua
+./test_scripts/Security/GetSystemTime/024_GetSystemTime_mobile_cer_valid_and_sdl_cer_becomes_not_valid.lua
+./test_scripts/Security/GetSystemTime/025_GetSystemTime_mobile_cer_and_sdl_cer_become_expired.lua
+./test_scripts/Security/GetSystemTime/026_GetSystemTime_mobile_cer_and_sdl_cer_become_not_valid_yet.lua
+./test_scripts/Security/GetSystemTime/027_GetSystemTime_without_response_from_HMI.lua
